requirements.txt
asgiref==3.3.1
decorator==4.4.2
Django==3.1.4
Pillow==8.0.1
pytz==2020.4
self==2020.12.3
sqlparse==0.4.1
